requirements.txt
a2wsgi==1.10.7
aiodns==3.2.0
aioquic==1.2.0
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
checks==0.2
ConfigParser==7.1.0
contextlib2==21.6.0
curio==1.6
cx_Oracle==8.3.0
Cython==3.0.11
datasets==3.2.0
defusedxml==0.7.1
docutils==0.21.2
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.16.1
grandalf==0.8
gunicorn==23.0.0
h2==4.1.0
HTMLParser==0.0.2
hypothesis==6.122.6
importlib_metadata==8.5.0
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jnius==1.1.0
js==1.0
jsonschema==4.23.0
keyring==25.5.0
langchain_anthropic==0.3.1
langchain_astradb==0.5.2
langchain_aws==0.2.10
langchain_chroma==0.1.4
langchain_cohere==0.3.4
langchain_community==0.3.13
langchain_elasticsearch==0.3.0
langchain_experimental==0.3.4
langchain_fireworks==0.2.6
langchain_google_genai==2.0.7
langchain_google_vertexai==2.0.9
langchain_groq==0.2.2
langchain_huggingface==0.1.2
langchain_mistralai==0.2.4
langchain_mongodb==0.3.0
langchain_neo4j==0.2.0
langchain_openai==0.2.14
langchain_pinecone==0.2.0
langchain_postgres==0.0.12
langchain_qdrant==0.2.0
langchain_together==0.2.0
langchain_weaviate==0.0.3
langchainhub==0.1.21
langsmith_pyo3==0.1.0rc5
lark==1.2.2
lxml==5.3.0
mask==1.0.0
mtrand==0.1
numexpr==2.10.2
objgraph==3.6.2
openapi_pydantic==0.5.0
outcome==1.3.0.post0
pickle5==0.0.12
Pillow==11.0.0
protobuf==5.29.2
psutil==6.1.1
pygraphviz==1.14
PyInstaller==6.11.1
pyodide==0.0.2
pyOpenSSL==24.3.0
pyppeteer==2.0.0
pysqlcipher3==1.2.0
pytest==8.3.4
pytz==2024.2
pywin32==308
rapidfuzz==3.11.0
scipy_doctest==1.6
setuptools==75.6.0
simplejson==3.19.3
simsimd==6.2.1
socksio==1.0.0
speedups==1.4.0
Sphinx==8.1.3
sqlcipher3==0.5.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
ujson==5.10.0
url_parser==3.0.3
urllib3_secure_extra==0.1.0
wmi==1.5.1
wsproto==1.2.0
xmlrpclib==1.0.1
zstandard==0.23.0
